refactor: simplify cmd
