--- src/cmd/ing/refresh-ing-list.ts
@@ -1,3 +1,0 @@
-import { getIngListWebviewProvider } from '@/service/ing/ing-list-webview-provider'
-
-export const refreshIngList = () => getIngListWebviewProvider().refreshingList()
