--- src/cmd/open/open-cnb-ing.ts
@@ -1,4 +1,0 @@
-import { Uri } from 'vscode'
-import { execCmd } from '@/infra/cmd'
-
-export const openCnbIng = () => execCmd('vscode.open', Uri.parse('https://ing.cnblogs.com'))
