--- src/cmd/open/open-cnb-news.ts
@@ -1,4 +1,0 @@
-import { execCmd } from '@/infra/cmd'
-import { Uri } from 'vscode'
-
-export const openCnbNews = () => execCmd('vscode.open', Uri.parse('https://news.cnblogs.com'))
