--- src/cmd/open/open-my-blog-console.ts
@@ -1,4 +1,0 @@
-import { Uri } from 'vscode'
-import { execCmd } from '@/infra/cmd'
-
-export const openMyWebBlogConsole = () => execCmd('vscode.open', Uri.parse('https://i.cnblogs.com'))
