--- src/cmd/open/os-open-workspace.ts
@@ -1,4 +1,0 @@
-import { execCmd } from '@/infra/cmd'
-import { WorkspaceCfg } from '@/ctx/cfg/workspace'
-
-export const osOpenWorkspace = () => execCmd('revealFileInOS', WorkspaceCfg.getWorkspaceUri())
