+++ src/model/blog-export/export-post.ts
@@ -1,16 +1,14 @@
-import { Post } from '@/model/post'
 import { Model } from 'sequelize'
+import { Post } from '@/model/post'
 
-export type PostType = 'BlogPost' | 'Message'
+export type ExportPostType = 'BlogPost' | 'Message'
 
 export type ExportPost = Pick<
     Post,
     'id' | 'title' | 'datePublished' | 'blogId' | 'isMarkdown' | 'accessPermission' | 'entryName' | 'dateUpdated'
 > & {
     body?: string | null
-    postType: PostType
+    postType: ExportPostType
 }
 
 export class ExportPostModel extends Model<ExportPost> {}
-
-export { Post } from '@/model/post'
