+++ src/model/post-cfg.ts
@@ -1,6 +1,4 @@
-import { Post } from './post'
+import { Post } from '@/model/post'
 
-type PostCfg = Partial<Omit<Post, 'postBody' | 'id' | 'author' | 'blogId' | 'blogTeamIds'>> &
+export type PostCfg = Partial<Omit<Post, 'postBody' | 'id' | 'author' | 'blogId' | 'blogTeamIds'>> &
     Required<{ [P in keyof Post as Post[P] extends boolean ? P : never]: Post[P] }>
-
-export { PostCfg }
